requirements.txt
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
commonmark==0.9.1
idna==3.4
inquirerpy==0.3.4
pfzy==0.3.4
prompt-toolkit==3.0.33
Pygments==2.13.0
python-dotenv==0.21.0
requests==2.28.1
rich==12.6.0
typer==0.7.0
urllib3==1.26.13
wcwidth==0.2.5
